commit new changes of my version

+++ src/main/scala/IO/HTMLGenerator.scala
@@ -16,7 +16,7 @@ object HTMLGenerator {
     bw.close()
   }
 
-  def outPut(resultSet:ArrayBuffer[(String,Vector[Float])], queryPoint:(String,Vector[Float]), path:String) : Unit = {
+  def outPut(resultSet:ArrayBuffer[(String,Vector[Double])], queryPoint:(String,Vector[Double]), path:String) : Unit = {
     val sb = new StringBuilder
     // HEADER
     sb.append("<!doctype html>")
