+++ src/main/scala/IO/Parser.scala
@@ -35,8 +35,8 @@ class Parser(data:File) extends Serializable  {
 
   def hasNext : Boolean = iterator.hasNext
 
-  def next : (String, Vector[Float]) = {
+  def next : (String, Vector[Double]) = {
     val set = iterator.take(2).toList
-    (set(0).toString.substring(49), set(1).toString.split(" ").map(x=>x.toFloat).toVector)
+    (set(0).toString.substring(49), set(1).toString.split(" ").map(x=>x.toDouble).toVector)
   }
 }
