+++ src/main/scala/main/Query.scala
@@ -64,7 +64,7 @@ object Query {
         println(" loaded!")
         val parser = new Parser(config.queries)
         var i = 0
-        var queryPoints = new ArrayBuffer[(String, Vector[Float])](parser.size)
+        var queryPoints = new ArrayBuffer[(String, Vector[Double])](parser.size)
         while(i < parser.size) {
           i += 1
           queryPoints += parser.next
